courses/computer-security/project/pgp/PGP/requirements.txt
pycryptodome
rsa
pyqt5